requirements.txt
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
requests==2.32.3
soupsieve==2.6
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.1.3
